Add rates

+++ db/migrate/20220916185104_create_rates.rb
@@ -1,7 +1,7 @@
 class CreateRates < ActiveRecord::Migration[7.0]
   def change
     create_table :rates do |t|
-      t.decimal :rate
+      t.float :rate
 
       t.timestamps
     end
